Default chat mode to short + show real error type for save failures
Two small follow-ups from the writer's recent testing:

Default to short chat mode
- AIAssistantView's @State generationMode was initialized to .long
  (multi-pass long-form). The writer wanted quick chat by default
  (the long-form path triggers the chapter-write intent + multi-pass
  generation, which is heavy for a quick question).
- Changed to .short so opening the AI panel gives a fast chat model.
  The user can still click 'Long form' to switch.

Real error type for non-BackendError save failures
- The 'Save failed: HTTP 0: ...' toast was a placeholder for any
  error that wasn't a BackendError. URLError (network, timeout, host
  unreachable) was the most common culprit, but the user just saw
  'HTTP 0' with no clue what was actually wrong.
- Three new catch arms in saveNow:
    1. CancellationError — silent (expected when user switches
       chapters mid-save). Just log, don't toast.
    2. URLError — surface the rawValue code (-1001 timeout,
       -1004 can't connect, -1005 lost, etc.) plus the description.
       User sees e.g. 'Save failed: URLError -1004: Could not
       connect to host'.
    3. Any other Error — wraps as BackendError.networkError, prints
       the type name to stderr so we can debug from the logs.
- Expanded isNetworkError to recognize URLError messages so the
  5s auto-retry actually fires when the backend is restarting.

All 348 tests still pass.

+++ frontend/Quill/Views/AIAssistantView.swift
@@ -15,7 +15,7 @@ struct AIAssistantView: View {
     var width: CGFloat = 0
 
     @State private var inputText: String = ""
-    @State private var generationMode: AppState.GenerationMode = .long
+    @State private var generationMode: AppState.GenerationMode = .short
     @State private var outlineHint: String = ""
     @State private var chatFocusToken: Int = 0
     @ObservedObject private var slotRegistry = LLMSlotRegistry.shared
